client/requirements.txt
Flask>=2.2.1,<2.3
Flask-SocketIO>=5.3.2
requests>=2.27.0,<3
websocket-client>=1.7
numpy>=1.24.0